+++ Main/Domain/Interfaces/IUser.cs
@@ -11,7 +11,7 @@ namespace _3D_Tim_backend.Domain
         int LoseCountInRoom { get; set; }
         int TieCountInRoom { get; set; }
         int TotalBetsInRoom { get; set; }
-        int MoneyInRoom { get; set; }
-        ConcurrentDictionary<string, int> BetSides { get; set; }
+        long MoneyInRoom { get; set; }
+        ConcurrentDictionary<string, long> BetSides { get; set; }
     }
 }
