+++ Main/Entities/User.cs
@@ -20,9 +20,9 @@ namespace _3D_Tim_backend.Entities
         [StringLength(1000)]
         public string? PasswordHash { get; set; }
 
-        public int Money { get; set; } = 100000;
+        public long Money { get; set; } = 100000;
 
-        public int TotalBets { get; set; } = 0;
+        public long TotalBets { get; set; } = 0;
 
         public Role Role { get; set; } = Role.Guest;
